requirements.txt
snowflake-connector-python==3.6.0
apache-airflow-providers-snowflake==5.1.2
